fix 'multiple definition' errors
